+++ src/contexts/ServiceContext.js
@@ -220,164 +220,6 @@ export const ServiceProvider = ({ children }) => {
 
   /**
    *
-   * ██████╗░██╗░░░██╗██████╗░██╗░░░░░██╗░█████╗░
-   * ██╔══██╗██║░░░██║██╔══██╗██║░░░░░██║██╔══██╗
-   * ██████╔╝██║░░░██║██████╦╝██║░░░░░██║██║░░╚═╝
-   * ██╔═══╝░██║░░░██║██╔══██╗██║░░░░░██║██║░░██╗
-   * ██║░░░░░╚██████╔╝██████╦╝███████╗██║╚█████╔╝
-   * ╚═╝░░░░░░╚═════╝░╚═════╝░╚══════╝╚═╝░╚════╝░
-   *
-   * ██████╗░███████╗░██████╗░█████╗░██╗░░░██╗██████╗░░█████╗░███████╗░██████╗
-   * ██╔══██╗██╔════╝██╔════╝██╔══██╗██║░░░██║██╔══██╗██╔══██╗██╔════╝██╔════╝
-   * ██████╔╝█████╗░░╚█████╗░██║░░██║██║░░░██║██████╔╝██║░░╚═╝█████╗░░╚█████╗░
-   * ██╔══██╗██╔══╝░░░╚═══██╗██║░░██║██║░░░██║██╔══██╗██║░░██╗██╔══╝░░░╚═══██╗
-   * ██║░░██║███████╗██████╔╝╚█████╔╝╚██████╔╝██║░░██║╚█████╔╝███████╗██████╔╝
-   * ╚═╝░░╚═╝╚══════╝╚═════╝░░╚════╝░░╚═════╝░╚═╝░░╚═╝░╚════╝░╚══════╝╚═════╝░
-   */
-
-  /**
-   * Function to get the names of the provinces
-   * @async
-   * @param {string} province - The name of the province(can be misspelled), it will be searched in the georef api
-   * @returns {Promise<Array>}
-   */
-  const get_province_names = async (province) => {
-    try {
-      const response = await axios.get(`${URL_GEOREF}/provincias`, {
-        params: {
-          nombre: province,
-          campos: "estandar",
-          aplanar: true,
-        },
-      });
-      const { cantidad, provincias } = response.data;
-      return cantidad > 0
-        ? provincias.map((elemento) => elemento.nombre).sort()
-        : [];
-    } catch (error) {
-      console.error("Error al obtener las provincias:", error);
-      return [];
-    }
-  };
-
-  /**
-   * Asynchronously fetches city names based on province and city names.
-   * @async
-   * @param {string} province - the name of the province (can't be misspelled)
-   * @param {string} city - the name of the city in the province (can be misspelled)
-   * @return {Array<string>} an array of city names with department names, or an empty array
-   */
-  const get_citys_name = async (province, city) => {
-    try {
-      const response = await axios.get(`${URL_GEOREF}/localidades`, {
-        params: {
-          provincia: province,
-          nombre: city,
-          campos: "estandar", // 'estandar' or 'basico'
-          aplanar: true,
-        },
-      });
-      const { cantidad, localidades } = response.data;
-      return cantidad > 0
-        ? localidades.map(
-            (elemento) => `${elemento.nombre},${elemento.departamento_nombre}`
-          )
-        : [];
-    } catch (error) {
-      console.error("Error al obtener las localidades:", error);
-      return [];
-    }
-  };
-
-  /**
-   * Asynchronously retrieves address names based on the given province, city, and address.
-   * @async
-   * @param {string} province - The province where the address is located (can't be misspelled).
-   * @param {string} city - The city where the address is located (can't be misspelled).
-   * @param {string} address - The address to be searched (can be misspelled).
-   * @returns {Array<string>} - An array of formatted address names.
-   */
-  const get_address_names = async (province, city, address) => {
-    try {
-      // Split the city into locality and department
-      const [locality, department] = city.split(",");
-      // Make a GET request to retrieve address names
-      const response = await axios.get(`${URL_GEOREF}/direcciones`, {
-        params: {
-          direccion: address,
-          provincia: province,
-          departamento: department,
-          localidad: locality,
-          campos: "basico",
-          aplanar: true,
-        },
-      });
-      const { cantidad, direcciones } = response.data;
-      // Process the retrieved address names
-      return cantidad > 0
-        ? direcciones
-            .map((elemento) => {
-              // Format the street name and house number
-              const calleNombre = `${elemento.calle_nombre
-                .charAt(0)
-                .toUpperCase()}${elemento.calle_nombre.slice(1).toLowerCase()}`;
-              const alturaValor = elemento.altura_valor
-                ? elemento.altura_valor
-                : "NUMERO";
-              return `${calleNombre}, ${alturaValor}`;
-            })
-            .filter((elemento, index, self) => index === self.indexOf(elemento)) // Remove duplicate addresses
-            .sort() // Sort the addresses
-        : [];
-    } catch (error) {
-      console.error("Error al obtener las direcciones:", error);
-      return [];
-    }
-  };
-
-  /**
-   * Fetches the news list from the backend API.
-   * @async
-   * @return {Promise<Array>} the response data from the backend API
-   */
-  const fetch_news_list = async () => {
-    try {
-      const response = await axios.post(
-        `${URL_BACKEND}/api/resources/get-news-list`,
-        {},
-        { headers: { "X-API-Key": APP_KEY } }
-      );
-      return response.data;
-    } catch (error) {
-      console.error("Error fetching news list:", error);
-      return null;
-    }
-  };
-
-  /**
-   * Fetches the news PDF for a given ID.
-   *
-   * @param {type} id - The ID of the news PDF to fetch
-   * @return {Promise<string>} The url of the PDF
-   */
-  const fetch_news_pdf = async (id) => {
-    try {
-      const response = await axios.post(
-        `${URL_BACKEND}/api/resources/get-news-pdf`,
-        {
-          id: id,
-        },
-        { headers: { "X-API-Key": APP_KEY } }
-      );
-      return response.data;
-    } catch (error) {
-      console.error("Error fetching PDF viewer:", error);
-      return null;
-    }
-  };
-
-  /**
-   *
    * ██╗░░░██╗░██████╗███████╗██████╗░  ██████╗░███████╗░██████╗░█████╗░██╗░░░██╗██████╗░░█████╗░███████╗░██████╗
    * ██║░░░██║██╔════╝██╔════╝██╔══██╗  ██╔══██╗██╔════╝██╔════╝██╔══██╗██║░░░██║██╔══██╗██╔══██╗██╔════╝██╔════╝
    * ██║░░░██║╚█████╗░█████╗░░██████╔╝  ██████╔╝█████╗░░╚█████╗░██║░░██║██║░░░██║██████╔╝██║░░╚═╝█████╗░░╚█████╗░
@@ -530,119 +372,6 @@ export const ServiceProvider = ({ children }) => {
 
   /**
    *
-   * ███████╗███╗░░░███╗░█████╗░██╗██╗░░░░░  ██╗░░░██╗███████╗██████╗░██╗███████╗██╗░░░██╗
-   * ██╔════╝████╗░████║██╔══██╗██║██║░░░░░  ██║░░░██║██╔════╝██╔══██╗██║██╔════╝╚██╗░██╔╝
-   * █████╗░░██╔████╔██║███████║██║██║░░░░░  ╚██╗░██╔╝█████╗░░██████╔╝██║█████╗░░░╚████╔╝░
-   * ██╔══╝░░██║╚██╔╝██║██╔══██║██║██║░░░░░  ░╚████╔╝░██╔══╝░░██╔══██╗██║██╔══╝░░░░╚██╔╝░░
-   * ███████╗██║░╚═╝░██║██║░░██║██║███████╗  ░░╚██╔╝░░███████╗██║░░██║██║██║░░░░░░░░██║░░░
-   * ╚══════╝╚═╝░░░░░╚═╝╚═╝░░╚═╝╚═╝╚══════╝  ░░░╚═╝░░░╚══════╝╚═╝░░╚═╝╚═╝╚═╝░░░░░░░░╚═╝░░░
-   */
-
-  /**
-   * Re-sends the verification email to the user if they haven't verified their email address yet.
-   * @async
-   * @returns {boolean} true if the verification email was sent successfully, false otherwise. True don't means that the email was already verified.
-   * @throws {Error} if there was an error sending the verification email.
-   */
-  const resend_verify_email = async () => {
-    try {
-      const response = await axios.post(
-        `${URL_BACKEND}/api/email/verification-notification`, // Route to send the verification notification email
-        {}, // Empty data object
-        { headers: { "X-API-Key": APP_KEY } } // Request headers
-      );
-      return true; // Return true if the request was successful
-    } catch (error) {
-      console.error("Error al reenviar el correo:", error); // Log the error
-      return false; // Return false if there was an error
-    }
-  };
-
-  /**
-   * Sends the credentials to verify the email, and returns true if the email is verified.
-   * This function is called when the user clicks on the verification link in the email.
-   * @async
-   * @param {string} id - the ID for verification
-   * @param {string} hash - the hash for verification
-   * @param {number} expires - the expiration time for the verification link
-   * @param {string} signature - the signature for verification
-   * @return {boolean} true if email is verified, false if not
-   */
-  const send_confirmation_verify = async (id, hash, expires, signature) => {
-    try {
-      const response = await axios.post(
-        `${URL_BACKEND}/api/email/verify/${id}/${hash}?expires=${expires}&signature=${signature}`,
-        {},
-        { headers: { "X-API-Key": APP_KEY } }
-      );
-      const { message } = response.data;
-      return (
-        message === "Email verified" || message === "Email already verified"
-      );
-    } catch (error) {
-      console.error("Error during email verification:", error);
-      return false;
-    }
-  };
-
-  /**
-   * Sends confirmation code to the specified email address for verification
-   * @async
-   * @param {string} code - The confirmation code to send
-   * @param {string} email - The email address to send the confirmation code to
-   * @returns {boolean} - True if the confirmation code was successfully sent, otherwise false
-   */
-  const send_confirmation_code = async (code, email) => {
-    try {
-      // Send a POST request to the backend API to confirm the email verification
-      const response = await axios.post(
-        `${URL_BACKEND}/api/auth/email/verify/confirm`,
-        {
-          code: code,
-          email: email,
-        },
-        { headers: { "X-API-Key": APP_KEY } }
-      );
-      const { message } = response.data;
-      if (message === "Confirmation successful") {
-        return true;
-      }
-      return false;
-    } catch (error) {
-      // Log an error message if there was an error during email verification
-      console.error("Error during email verification:", error);
-      return false;
-    }
-  };
-
-  /**
-   * Send confirmation email
-   * @param {string} password - The password to use for verification
-   * @param {string} email - The email address to send the confirmation email to
-   * @returns {Promise<boolean>} - True if confirmation is successful, otherwise false
-   */
-  const send_confirmation_email = async (password, email) => {
-    try {
-      // Send a post request to the backend API to send the confirmation email
-      const result = await axios.post(
-        `${URL_BACKEND}/api/auth/email/verify/send`,
-        {
-          password: password,
-          email: email,
-        },
-        { headers: { "X-API-Key": APP_KEY } }
-      );
-      const { message } = result.data;
-      return message === "Email sent";
-    } catch (error) {
-      // Log and return false if there's an error during the confirmation process
-      console.error("Error during code verification: ", error);
-      return false;
-    }
-  };
-
-  /**
-   *
    * ██████╗░░█████╗░░██████╗░██████╗░██╗░░░░░░░██╗░█████╗░██████╗░██████╗░  ███████╗░█████╗░██████╗░░██████╗░░█████╗░████████╗
    * ██╔══██╗██╔══██╗██╔════╝██╔════╝░██║░░██╗░░██║██╔══██╗██╔══██╗██╔══██╗  ██╔════╝██╔══██╗██╔══██╗██╔════╝░██╔══██╗╚══██╔══╝
    * ██████╔╝███████║╚█████╗░╚█████╗░░╚██╗████╗██╔╝██║░░██║██████╔╝██║░░██║  █████╗░░██║░░██║██████╔╝██║░░██╗░██║░░██║░░░██║░░░
@@ -752,8 +481,6 @@ export const ServiceProvider = ({ children }) => {
         isAuthenticated,
         setIsAuthenticated,
         User,
-        fetch_news_list,
-        fetch_news_pdf,
         setUser,
         send_forgot_password_email,
         serverDates,
@@ -765,13 +492,6 @@ export const ServiceProvider = ({ children }) => {
         unauthenticate,
         registerRequest,
         get_ae_dates,
-        send_confirmation_email,
-        send_confirmation_code,
-        get_province_names,
-        get_citys_name,
-        get_address_names,
-        send_confirmation_verify,
-        resend_verify_email,
         start_ae_n,
         AE,
         send_reset_password,
